Try and get the views tests into a good place.
